requirements.txt
-i https://pypi.org/simple
greenlet==3.2.1 ; python_version < '3.14' and platform_machine == 'aarch64' or (platform_machine == 'ppc64le' or (platform_machine == 'x86_64' or (platform_machine == 'amd64' or (platform_machine == 'AMD64' or (platform_machine == 'win32' or platform_machine == 'WIN32')))))
jinja2==3.1.6 ; python_version >= '3.7'
kaleido==0.2.1
linkify-it-py==2.0.3
markdown-it-py==3.0.0 ; python_version >= '3.8'
markupsafe==3.0.2 ; python_version >= '3.9'
mdit-py-plugins==0.4.2
mdurl==0.1.2 ; python_version >= '3.7'
memray==1.17.1
narwhals==1.38.0 ; python_version >= '3.8'
numpy==2.2.5 ; python_version == '3.11'
packaging==25.0 ; python_version >= '3.8'
pandas==2.2.3
platformdirs==4.3.7 ; python_version >= '3.9'
plotly==6.0.1
psycopg2-binary==2.9.10
pygments==2.19.1 ; python_version >= '3.8'
python-dateutil==2.9.0.post0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pytz==2025.2
rich==14.0.0 ; python_full_version >= '3.8.0'
six==1.17.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlalchemy==2.0.40
textual==3.2.0 ; python_full_version >= '3.8.1' and python_full_version < '4.0.0'
typing-extensions==4.13.2 ; python_version >= '3.8'
tzdata==2025.2 ; python_version >= '2'
uc-micro-py==1.0.3 ; python_version >= '3.7'
